finished backend side

+++ app.py
@@ -1,9 +1,9 @@
 '''Main App Controller'''
 
 from model.curso import Curso
-# from model.estudante import Estudante
+from model.estudante import Estudante
 from dao.cursoDAO import CursoDAO
-# from dao.estudanteDAO import EstudanteDAO
+from dao.estudanteDAO import EstudanteDAO
 
 # set FLASK_APP=app.py && flask run
 from flask import Flask, render_template as render, request, redirect, url_for
@@ -14,7 +14,7 @@ app = Flask(__name__)
 
 @app.route('/')
 def runing():
-    return "<h5>running</h5>"
+    return '<a href="/app-school/">running<a>'
 
 
 @app.route('/app-school')
@@ -23,6 +23,13 @@ def runing():
 @app.route('/app-school/index.html')
 def index():
     return render('index.html')
+
+
+# logradouro
+# numero
+# bairro
+# cidade
+# estado
 
 
 #######################################
@@ -43,9 +50,10 @@ def curso_novo():
 
 @app.route('/app-school/curso/inserir', methods=['POST'])
 def curso_inserir():
-    curso = Curso(request.form['nome'], request.form['sigla'])
-    CursoDAO().inserir(curso)
-    return redirect(url_for('curso_listar'))
+    if request.method == 'POST':
+        curso = Curso(request.form['nome'], request.form['sigla'])
+        CursoDAO().inserir(curso)
+        return redirect(url_for('curso_listar'))
 
 ##
 # verificao para saber se curso ja foi cadastrado
@@ -74,9 +82,10 @@ def curso_remover():
 
 @app.route('/app-school/curso/confirmar-exclusao', methods=['POST'])
 def curso_confirmar_exclusao():
-    curso = Curso(request.form.get('nome'), request.form['sigla'])
-    CursoDAO().remover(curso)
-    return redirect(url_for('curso_listar'))
+    if request.method == 'POST':
+        curso = Curso(request.form.get('nome'), request.form['sigla'])
+        CursoDAO().remover(curso)
+        return redirect(url_for('curso_listar'))
 
 
 #######################################
@@ -84,7 +93,65 @@ def curso_confirmar_exclusao():
 # .:ESTUDANTE                         #
 #                                     #
 #######################################
-# , estudantes=EstudanteDAO().listar())
+
 @app.route('/app-school/estudante/listar')
 def estudante_listar():
-    return render('/estudante/listar.html')
+    return render('/estudante/listar.html', estudantes=EstudanteDAO().listar())
+
+
+@app.route('/app-school/estudante/novo')
+def estudante_novo():
+    return render('/estudante/inserir.html', cursos=CursoDAO().listar())
+
+
+@app.route('/app-school/estudante/inserir', methods=['POST'])
+def estudante_inserir():
+    if request.method == 'POST':
+        estudante = Estudante(
+            request.form['nome'],
+            request.form['matricula'],
+            Curso(request.form['curso'], None))
+        EstudanteDAO().inserir(estudante)
+        return redirect(url_for('estudante_listar'))
+
+
+@app.route('/app-school/estudante/alterar', methods=['GET'])
+def estudante_alterar():
+    estudante = Estudante(None, request.args.get('matricula'), None)
+    return render('estudante/alterar.html',
+                  estudante=EstudanteDAO().listarPorID(estudante),
+                  cursos=CursoDAO().listar())
+
+
+@app.route('/app-school/estudante/confirmar-alteracao', methods=['POST'])
+def estudante_confirmar_alteracao():
+    if request.method == 'POST':
+        print(request.method)
+        print(request.form.get('nome'))
+        print(request.form.get('matricula'))
+        estudante = Estudante(
+            request.form.get('nome'),
+            request.form.get('matricula'),
+            Curso(request.form.get('curso'), None))
+        EstudanteDAO().alterar(estudante, request.form.get('matricula'))
+        return redirect(url_for('estudante_listar'))
+
+
+@app.route('/app-school/estudante/remover')
+def estudante_remover():
+    estudante = Estudante(None, request.args.get('matricula'), None)
+    return render('/estudante/remover.html',
+                  estudante=EstudanteDAO().listarPorID(estudante),
+                  cursos=CursoDAO().listar())
+
+
+@app.route('/app-school/estudante/confirmar-exclusao', methods=['POST'])
+def estudante_confirmar_exclusao():
+    if request.method == 'POST':
+        estudante = Estudante(
+            None,
+            request.form.get('matricula'),
+            None
+        )
+        EstudanteDAO().remover(estudante)
+        return redirect(url_for('estudante_listar'))
